+++ R/globalsOf.R
@@ -4,7 +4,7 @@
 #'
 #' @param envir The environment from where to search for globals.
 #'
-#' @param \dots Not used.
+#' @param \ldots Not used.
 #'
 #' @param method A character string specifying what type of search algorithm
 #'        to use.
@@ -60,7 +60,7 @@
 #' @example incl/globalsOf.R
 #'
 #' @seealso
-#' Internally, the \pkg{\link{codetools}} package is utilized for
+#' Internally, the \pkg{codetools} package is utilized for
 #' code inspections.
 #'
 #' @aliases findGlobals