+++ inst/testme/test-findGlobals.R
@@ -1,5 +1,4 @@
-source("incl/start.R")
-
+library(globals)
 
 message("*** findGlobals() ...")
 
@@ -61,6 +60,22 @@ assert_identical_sets(globals_i, c("{", "x"))
 globals_i <- findGlobals({ "x" <- 1; x }, substitute = TRUE)
 print(globals_i)
 assert_identical_sets(globals_i, c("{", "<-"))
+
+globals <- findGlobals(list())
+print(globals)
+assert_identical_sets(globals, character(0L))
+
+expr <- quote(list())
+attr(expr, "abc") <- quote({ a })
+attr(expr, "def") <- quote({ d })
+globals <- findGlobals(expr)
+print(globals)
+assert_identical_sets(globals, c("list", "{", "a", "d"))
+
+globals <- findGlobals(expr, attributes = "abc")
+print(globals)
+assert_identical_sets(globals, c("list", "{", "a"))
+
 
 
 message(" ** findGlobals(..., tweak):")
@@ -155,6 +170,10 @@ for (method in c("conservative", "liberal", "ordered")) {
   assert_identical_sets(globals_i, c("GLOBAL", "ARG"))
 }
 
+
+globals <- findGlobals(quote({ a * b }), trace = TRUE)
+print(globals)
+assert_identical_sets(globals, c("{", "*", "a", "b"))
+
 message("*** findGlobals() ... DONE")
 
-source("incl/end.R")